pyproject.toml
[project]
name = "dafne-dataset"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "datman",
    "numpy>=1.21.5",
    "pillow>=10.4.0",
    "requests>=2.32.4",
    "tqdm>=4.66.5",
]

[tool.uv.sources]
datman = { path = "../DatasetManager", editable = true }

[dependency-groups]
dev = [
    "beautifulsoup4==4.12.2",
]


